Added file menu functions to distNetViz.html. Modified a few routes in web.py to allow file menu functions to work. Began editing visual interface code in distNetViz.html. Added jasmine testing library to test front-end JavaScript code. Added tests for distNetViz.html in distNetVizSpec.js. Added a component file for waterheaters with waterheater.json. Modified cymeToGridlab.py so it won't fail to convert files, but it can still fail depending on the file it is converting.

+++ omf/models/storageDeferral.py
@@ -1,271 +1,156 @@
 ''' Calculate the costs and benefits of energy storage from a distribution utility perspective. '''
 
-import json, os, sys, tempfile, webbrowser, time, shutil, datetime, subprocess, traceback, csv, math, copy
-import multiprocessing
-import numpy as np
-from os.path import join as pJoin
-from  dateutil.parser import parse
-from numpy import npv
-from jinja2 import Template
+import sys, shutil, csv, math
+from os.path import isdir, join as pJoin
 from omf.models import __neoMetaModel__
 from __neoMetaModel__ import *
 
 # Model metadata:
 modelName, template = metadata(__file__)
-tooltip = "The storageDeferral model calculates the amount of energy storage capacity needed to reduce the load on a substation transformer or line below a user-defined limi"
-
-# # NOTE: used for debugging don't delete.
-# import matplotlib.pyplot as plt
+tooltip = ("The storageDeferral model calculates the amount of energy storage "
+	"capacity needed to reduce the load on a substation transformer or line "
+	"below a user-defined limit.")
 
 def work(modelDir, inputDict):
 	''' Run the model in a separate process. web.py calls this to run the model.
 	This function will return fast, but results take a while to hit the file system.'''
-	outData = {}
-	# Get variables.
-	cellCapacity = float(inputDict['cellCapacity'])
+	out = {}
 	(cellCapacity, dischargeRate, chargeRate, cellCost) = \
 		[float(inputDict[x]) for x in ('cellCapacity', 'dischargeRate', 'chargeRate', 'cellCost')]
-	battEff	= float(inputDict.get("batteryEfficiency", 92)) / 100.0 * float(inputDict.get("inverterEfficiency", 92)) / 100.0 * float(inputDict.get("inverterEfficiency", 92)) / 100.0
-	discountRate = float(inputDict.get('discountRate', 2.5)) / 100.0
-	dodFactor = float(inputDict.get('dodFactor', 85)) / 100.0
-	projYears = int(inputDict.get('projYears',10))
-	transformerThreshold = float(inputDict.get('transformerThreshold',6.5)) * 1000
-	negTransformerThreshold = transformerThreshold * -1
-	avoidedCost = int(inputDict.get('avoidedCost',2000000))
-	batteryCycleLife = int(inputDict.get('batteryCycleLife', 5000))
+	
 	deferralType = inputDict.get('deferralType')
-	retailCost = float(inputDict.get('retailCost', 0.06))
-	yearsToReplace = int(inputDict.get('yearsToReplace', 2))
-	carryingCost = float(inputDict.get('carryingCost', 10))/100
-	# Put demand data in to a file for safe keeping.
-	with open(pJoin(modelDir,"demand.csv"),"w") as demandFile:
-		demandFile.write(inputDict['demandCurve'])
-	# Start running battery simulation.
-	battCapacity = cellCapacity * dodFactor
-	battDischarge = dischargeRate
-	battCharge = chargeRate
-	# Most of our data goes inside the dc "table"
-	try:
-		dc = []
-		with open(pJoin(modelDir,"demand.csv")) as inFile:
-			reader = csv.reader(inFile)
-			for row in reader:
-				dc.append({'power': float(row[0])})
-			if len(dc)!=8760: raise Exception
+	projYears = int(inputDict.get('yearsToReplace'))
+	avoidedCost = int(inputDict.get('avoidedCost'))
+	yearsToReplace = int(inputDict.get('yearsToReplace'))
+	batteryCycleLife = int(inputDict.get('batteryCycleLife'))
+	retailCost = float(inputDict.get('retailCost'))
+	dodFactor = float(inputDict.get('dodFactor')) / 100.0
+	carryingCost = float(inputDict.get('carryingCost')) / 100.0
+	
+	threshold = float(inputDict.get('transformerThreshold')) * 1000
+	inverterEfficiency = float(inputDict.get("inverterEfficiency", 92)) / 100.0
+	# NOTE: Inverter Efficiency is round-trip
+	battEff = float(inputDict.get("batteryEfficiency", 92)) / 100.0 * inverterEfficiency ** 2
+	
+	with open(pJoin(modelDir,'demand.csv'),'w') as f:
+		f.write(inputDict['demandCurve'])
+	
+	dc = []
+	try:	
+		with open(pJoin(modelDir,'demand.csv')) as f:
+			for r in csv.reader(f):
+				dc.append({	'power': abs(float(r[0])),
+							'excessDemand': abs(float(r[0])) - threshold,
+							'negative': True if float(r[0]) < 0 else False })
+			assert len(dc) == 8760
 	except:
-			e = sys.exc_info()[0]
-			if str(e) == "<type 'exceptions.SystemExit'>":
-				pass
+		if str(sys.exc_info()[0]) != "<type 'exceptions.SystemExit'>":
+			raise Exception("Demand CSV file is incorrect format.")
+	
+	'''
+	Find when demand surpasses the threshold. Sum the excessDemand until 
+	its negative component can appropriately charge a battery that 
+	supports its positive component. That is one group. Sum the positive
+	excessDemand for that group. Use the group with the largest positive
+	sum to calculate how many batteries to use. 
+	'''
+	inloop, excess, excessDemandArray = False, [], []
+	for r in dc:
+		if not inloop:
+			# Enter loop
+			if r['excessDemand'] > 0:
+				excess.append(r['excessDemand'])
+				inloop = True
+		else:
+			# in loop
+			if sum(excess) > 0:
+				excess.append(r['excessDemand'])
+			# exit loop
 			else:
-				errorMessage = "Demand CSV file is incorrect format."
-				raise Exception(errorMessage)
-	if deferralType == 'subTransformer':
-		for row in dc:
-			row['excessDemand'] = row['power'] - transformerThreshold
-		excessDemand = 0
-		rechargePotential = 0
-		excessDemandArray = []
-		#Determine the max excess demand/energy
-		for row in dc:
-			if row['excessDemand'] > 0:
-				excessDemand += row['excessDemand']
-				if rechargePotential < 0:	
-					excessDemandArray.append(rechargePotential)
-				rechargePotential = 0
-			elif row['excessDemand'] < 0:
-				rechargePotential += row['excessDemand']
-				if excessDemand > 0:
-					excessDemandArray.append(excessDemand)
-				excessDemand = 0
-		finalDC = copy.deepcopy(dc)
-		try:
-			excessDemandMax = max(excessDemandArray)
-		except:
-			errorMessage = "Demand Curve does not exceed Threshold Capacity. Lower Threshold Capacity and run again."
-			raise Exception(errorMessage)
-		#Calculate the number of units needed to cover the max demand/energy
-		numOfUnits = math.ceil(excessDemandMax/(cellCapacity))
-		newBattCapacity = numOfUnits *cellCapacity * dodFactor
-		newBattDischarge = numOfUnits * dischargeRate
-		newBattCharge = numOfUnits * chargeRate
-		run = True
-		counter = 0
-		#Iterate through the demand curve, if the battery runs out of charge and there is still excess demand: add more batteries
-		while run == True:
-			newBattCapacity = numOfUnits *cellCapacity * dodFactor
-			battSoC = newBattCapacity
-			for row in dc:
-				# month = int(row['datetime'].month)-1
-				discharge = min(numOfUnits * dischargeRate,battSoC, row['power'] - transformerThreshold)
-				charge = min(numOfUnits * chargeRate, newBattCapacity-battSoC, (transformerThreshold - row['power'])*battEff)
-				if counter == 8760:
-					run = False
-				else:
-					if row['power'] > transformerThreshold and battSoC > (row['power'] - transformerThreshold):
-						battSoC -= discharge
-						counter += 1
-					elif row['power'] <= transformerThreshold and battSoC < newBattCapacity:
-						battSoC += charge
-						counter += 1
-					elif row['power'] > transformerThreshold and battSoC < (row['power'] - transformerThreshold):
-						numOfUnits += math.ceil((row['power']-transformerThreshold-battSoC)/cellCapacity)
-						counter = 0
-						break
-					else:
-						counter += 1
-		afterBattCapacity = numOfUnits *cellCapacity * dodFactor
-		afterBattDischarge = numOfUnits * dischargeRate
-		afterBattCharge = numOfUnits * chargeRate
-		battSoC = afterBattCapacity
-		#Battery dispatch loop, determine net power, battery state of charge.
-		for row in finalDC:
-			outData['startDate'] = '2011-01-01' #finalDC[0]['datetime'].isoformat()
-			# month = int(row['datetime'].month)-1
-			discharge = min(afterBattDischarge,battSoC, row['power'] - transformerThreshold)
-			charge = min(afterBattCharge, afterBattCapacity-battSoC, (transformerThreshold - row['power'])*battEff)
-			if row['power'] > transformerThreshold and battSoC > 0:
-				row['netpower'] = row['power'] - discharge
-				battSoC -= discharge
-			elif row['power'] <= transformerThreshold and battSoC < afterBattCapacity:
-				row['netpower'] = row['power'] + charge/battEff
-				battSoC += charge
-			else:
-				row['netpower'] = row['power']
-			row['battSoC'] = battSoC
-	else: # Line deferral option.
-		for row in dc:
-			if row['power']>0:
-				row['excessDemand'] = row['power'] - transformerThreshold
-				row['sign'] = 'positive'
-			else:
-				row['excessDemand'] = row['power'] - negTransformerThreshold
-				row['sign'] = 'negative'
-		excessDemand = 0
-		rechargePotential = 0
-		excessDemandArray = []
-		for row in dc:
-			if (row['sign']=='positive' and row['power'] - transformerThreshold < 0) or (row['sign']=='negative' and row['excessDemand'] > 0):
-				if excessDemand > 0:
-					excessDemandArray.append(excessDemand)
-				excessDemand = 0
-				if row['excessDemand'] > 0:
-					row['excessDemand'] = -1 * row['excessDemand']
-				rechargePotential += row['excessDemand']
-			elif (row['sign']=='positive' and row['power'] - transformerThreshold > 0) or (row['sign']=='negative' and row['excessDemand'] < 0):
-				if rechargePotential < 0:	
-					excessDemandArray.append(rechargePotential)
-				rechargePotential = 0
-				excessDemand += abs(row['excessDemand'])
-		try:
-			excessDemandMax = max(excessDemandArray)
-		except:
-			errorMessage = "Demand Curve does not exceed Threshold Capacity. Lower Threshold Capacity and run again."
-			raise Exception(errorMessage)
-		numOfUnits = math.ceil(excessDemandMax/(cellCapacity))
-		finalDC = copy.deepcopy(dc)
-		newBattCapacity = numOfUnits *cellCapacity * dodFactor
-		newBattDischarge = numOfUnits * dischargeRate
-		newBattCharge = numOfUnits * chargeRate
-		run = True
-		counter = 0
-		while run == True:
-			newBattCapacity = numOfUnits *cellCapacity * dodFactor
-			battSoC = newBattCapacity
-			for row in dc:
-				# month = int(row['datetime'].month)-1
-				if row['sign'] == 'positive':
-					discharge = min(numOfUnits * dischargeRate,battSoC, row['power'] - transformerThreshold)
-					charge = min(numOfUnits * chargeRate, newBattCapacity-battSoC, (transformerThreshold - row['power'])*battEff)
-				elif row['sign'] == 'negative':
-					discharge = min(numOfUnits * dischargeRate,battSoC, abs(row['power']) - transformerThreshold)
-					charge = min(numOfUnits * chargeRate, newBattCapacity-battSoC, (transformerThreshold - abs(row['power']))*battEff)
-				if counter == 8760:
-					run = False
-				else:
-					if abs(row['power']) > transformerThreshold and battSoC > (abs(row['power']) - transformerThreshold):
-						battSoC -= abs(discharge)
-						counter += 1
-					elif abs(row['power']) < transformerThreshold and battSoC < newBattCapacity:
-						battSoC += abs(charge)
-						counter += 1
-					elif abs(row['power']) > transformerThreshold and battSoC < (abs(row['power']) - transformerThreshold):
-						numOfUnits = numOfUnits + math.ceil((abs(row['power'])-transformerThreshold-battSoC)/cellCapacity)
-						counter = 0
-						break
-					else:
-						counter += 1
-		afterBattCapacity = numOfUnits *cellCapacity * dodFactor
-		afterBattDischarge = numOfUnits * dischargeRate
-		afterBattCharge = numOfUnits * chargeRate
-		battSoC = afterBattCapacity
-		finalDC = copy.deepcopy(dc)
-		for row in finalDC:
-			outData['startDate'] = '2011-01-01'#finalDC[0]['datetime'].isoformat()
-			# month = int(row['datetime'].month)-1
-			discharge = abs(row['power']) - transformerThreshold
-			charge = min(afterBattCharge, afterBattCapacity-battSoC, battEff*(transformerThreshold - abs(row['power'])))
-			if (row['sign'] == 'positive' and row['excessDemand'] > 0) and battSoC > 0:
-				row['netpower'] = row['power'] - discharge
-				battSoC -= abs(discharge)
-			elif (row['sign'] == 'negative' and row['power'] + transformerThreshold < 0 ) and battSoC > 0:
-				row['netpower'] = row['power'] + discharge
-				battSoC -= abs(discharge)
-			elif (row['sign'] == 'positive' and row['excessDemand'] < 0) and battSoC < afterBattCapacity:
-				row['netpower'] = row['power'] + charge/battEff
-				battSoC += abs(charge)
-			elif (row['sign'] == 'negative' and row['power'] + transformerThreshold > 0 ) and battSoC < afterBattCapacity:
-				row['netpower'] = row['power'] - charge/battEff
-				battSoC += abs(charge)
-			else:
-				row['netpower'] = row['power']
-			row['battSoC'] = battSoC
-	#Calculations
-	yearlyCarryCost = carryingCost * avoidedCost
-	carryCost = yearlyCarryCost * yearsToReplace
-	outData['numOfBatteries'] = numOfUnits
-	outData['batteryCost'] = numOfUnits * cellCost
-	outData['avoidedCost'] = carryCost
-	outData['netAvoidedCost'] = carryCost - (numOfUnits * cellCost)
-	outData['demand'] = [t['power']*1000.0 for t in finalDC]
-	outData['demandAfterBattery'] = [t['netpower']*1000.0 for t in finalDC]
-	demandAfterBattery = outData['demandAfterBattery']
-	demand = outData['demand']
-	outData['batteryDischargekW'] = [abs(demand) - abs(demandAfterBattery) for demand, demandAfterBattery in zip(demand, demandAfterBattery)]
-	outData['batteryDischargekWMax'] = max(outData['batteryDischargekW'])
-	outData['transformerThreshold'] = transformerThreshold * 1000.0
-	outData['batterySoc'] = [t['battSoC']/newBattCapacity*100.0*dodFactor + (100-100*dodFactor) for t in finalDC]
-	SoC = outData['batterySoc']
-	cycleEquivalents = sum([SoC[i]-SoC[i+1] for i,x in enumerate(SoC[0:-1]) if SoC[i+1] < SoC[i]]) / 100.0
-	outData['cycleEquivalents'] = cycleEquivalents
-	outData['batteryLife'] = batteryCycleLife/cycleEquivalents
-	battCostPerCycle =  afterBattCapacity * cellCost / batteryCycleLife
-	lcoeTotCost = (cycleEquivalents *  afterBattCapacity * retailCost) + (battCostPerCycle * cycleEquivalents)
-	loceTotEnergy = cycleEquivalents * afterBattCapacity
-	LCOE = lcoeTotCost / loceTotEnergy
-	outData['LCOE'] = LCOE
-	# Stdout/stderr.
-	outData["stdout"] = "Success"
-	outData["stderr"] = ""
-	return outData
+				excessDemandArray.append(sum([i for i in excess if i > 0]))
+				inloop, excess = False, []
+	assert len(excessDemandArray) != 0, ("Demand Curve does not exceed "
+		"Threshold Capacity. Lower Threshold Capacity and run again.")	
+	# High threshold
+	numOfUnits = math.ceil(max(excessDemandArray)/chargeRate)
+
+	found = False
+	while True:
+		battCapacity = numOfUnits * cellCapacity * dodFactor
+		SoC = battCapacity
+		for r in dc:
+			charge = (-1*min(numOfUnits*dischargeRate, SoC, r['excessDemand']) if r['power'] > threshold 
+				else min(numOfUnits*chargeRate, battCapacity-SoC, -1*r['excessDemand']))
+			r['netpower'] = r['power'] + charge
+			SoC += charge
+			r['battSoC'] = SoC
+
+		if found:
+			break
+
+		# reduce the 
+		# For particularly low thresholds this will take a while unfortunately.
+		if all([r['netpower'] <= threshold for r in dc]):
+			numOfUnits -= 1
+		else:
+			numOfUnits += 1
+			found = True
+
+	# Convert to negative where appropriate
+	for r in dc:
+		if r['negative']:
+			r['netpower'] *= -1
+			r['power'] *= -1
+	
+	# -------------------------- CALCULATIONS ------------------------ #	
+	
+	# Demand Before and After Storage Graph
+	out['demand'] = [t['power'] * 1000.0 for t in dc]
+	out['demandAfterBattery'] = [t['netpower'] * 1000.0 for t in dc]
+	out['batteryDischargekW'] = [abs(d) - abs(dab) for d, dab in zip(out['demand'], out['demandAfterBattery'])]
+	out['batteryDischargekWMax'] = max(out['batteryDischargekW'])
+	out['transformerThreshold'] = threshold * 1000.0 # kW -> W
+
+	# Battery SoC Graph
+	out['batterySoc'] = SoC = [t['battSoC']/battCapacity*100.0*dodFactor + (100-100*dodFactor) for t in dc]	
+	cycleEquivalents = sum([SoC[i]-SoC[i+1] for i, x in enumerate(SoC[:-1]) if SoC[i+1] < SoC[i]]) / 100.0
+	out['cycleEquivalents'] = cycleEquivalents
+	out['batteryLife'] = batteryCycleLife / cycleEquivalents
+	
+	# Avoided Cost Graph
+	out['numOfBatteries'] = numOfUnits
+	rechargeCost = sum([t for t in out['batteryDischargekW'] if t > 0])/battEff*retailCost
+	out['batteryCost'] = numOfUnits * cellCost + rechargeCost
+	out['avoidedCost'] = carryingCost * avoidedCost * yearsToReplace
+	out['netAvoidedCost'] = out['avoidedCost'] - (numOfUnits * cellCost)	
+	lcoeTotCost = cycleEquivalents*battCapacity*retailCost + cellCost/batteryCycleLife*cycleEquivalents
+	out['LCOE'] = lcoeTotCost / (cycleEquivalents * battCapacity)
+
+	# Other
+	out['startDate'] = '2011-01-01'
+	out["stdout"] = "Success"
+	out["stderr"] = ""
+
+	return out
 
 def new(modelDir):
 	''' Create a new instance of this model. Returns true on success, false on failure. '''
 	defaultInputs = {
 		"batteryEfficiency": "92",
 		"retailCost": "0.06",
-		"deferralType": "subTransformer",
 		"inverterEfficiency": "97.5",
 		"cellCapacity": "7",
 		"created": "2015-06-12 17:20:39.308239",
 		"dischargeRate": "5",
 		"modelType": modelName,
 		"chargeRate": "5",
-		"demandCurve": open(pJoin(__neoMetaModel__._omfDir,"static","testFiles","FrankScadaValidCSV_Copy.csv")).read(),
+		#"deferralType": "subTransformer", "demandCurve": open(pJoin(__neoMetaModel__._omfDir,"static","testFiles","FrankScadaValidCSV_Copy.csv")).read(),
+		"deferralType": "line", "demandCurve": open(pJoin(__neoMetaModel__._omfDir,"static","testFiles","negativeDemand.csv")).read(),
 		"fileName": "FrankScadaValidCSV_Copy.csv",
 		"cellCost": "7140",
 		"dodFactor":"100",
 		"avoidedCost":"2000000",
-		"transformerThreshold":"6.6",
+		"transformerThreshold":"6.3",
 		"batteryCycleLife": "5000",
 		"carryingCost":"7",
 		"yearsToReplace":"2"
@@ -273,22 +158,14 @@ def new(modelDir):
 	return __neoMetaModel__.new(modelDir, defaultInputs)
 
 def _tests():
-	# Location
 	modelLoc = pJoin(__neoMetaModel__._omfDir,"data","Model","admin","Automated Testing of " + modelName)
 	# Blow away old test results if necessary.
-	try:
+	if isdir(modelLoc):
 		shutil.rmtree(modelLoc)
-	except:
-		# No previous test results.
-		pass
-	# Create New.
-	new(modelLoc)
-	# Pre-run.
-	renderAndShow(modelLoc)
-	# Run the model.
-	runForeground(modelLoc)
-	# Show the output.
-	renderAndShow(modelLoc)
+	new(modelLoc) # Create New.
+	renderAndShow(modelLoc) # Pre-run.
+	runForeground(modelLoc) # Run the model.	
+	renderAndShow(modelLoc) # Show the output.
 
 if __name__ == '__main__':
 	_tests()
